+++ rotkehlchen/tests/unit/decoders/test_main.py
@@ -65,6 +65,7 @@ def test_decoders_initialization(ethereum_transaction_decoder: EthereumTransacti
         'HarvestFinance',
         'Juicebox',
         'Kyber',
+        'Lido',
         'Liquity',
         'Lockedgno',
         'Makerdao',
@@ -178,6 +179,7 @@ def test_decoders_initialization(ethereum_transaction_decoder: EthereumTransacti
         'shutter',
         'eigenlayer',
         'omni',
+        'lido',
     }